split enterprise_search files to read and payload
